# apps/web/frontend/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "PublicationColumnContractModel": {
        "additionalProperties": false,
        "properties": {
          "aggregation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aggregation"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "filterable": {
            "title": "Filterable",
            "type": "boolean"
          },
          "grain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grain"
          },
          "json_type": {
            "title": "Json Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nullable": {
            "title": "Nullable",
            "type": "boolean"
          },
          "semantic_role": {
            "title": "Semantic Role",
            "type": "string"
          },
          "sortable": {
            "title": "Sortable",
            "type": "boolean"
          },
          "storage_type": {
            "title": "Storage Type",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          }
        },
        "required": [
          "name",
          "storage_type",
          "json_type",
          "nullable",
          "description",
          "semantic_role",
          "filterable",
          "sortable"
        ],
        "title": "PublicationColumnContractModel",
        "type": "object"
      },
      "PublicationContractModel": {
        "additionalProperties": false,
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/PublicationColumnContractModel"
            },
            "title": "Columns",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "lineage_required": {
            "title": "Lineage Required",
            "type": "boolean"
          },
          "pack_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Name"
          },
          "pack_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pack Version"
          },
          "publication_key": {
            "title": "Publication Key",
            "type": "string"
          },
          "relation_name": {
            "title": "Relation Name",
            "type": "string"
          },
          "renderer_hints": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Renderer Hints",
            "type": "object"
          },
          "retention_policy": {
            "title": "Retention Policy",
            "type": "string"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "supported_renderers": {
            "items": {
              "type": "string"
            },
            "title": "Supported Renderers",
            "type": "array"
          },
          "ui_descriptor_keys": {
            "items": {
              "type": "string"
            },
            "title": "Ui Descriptor Keys",
            "type": "array"
          },
          "visibility": {
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "publication_key",
          "relation_name",
          "schema_name",
          "schema_version",
          "display_name",
          "visibility",
          "retention_policy",
          "lineage_required",
          "supported_renderers",
          "renderer_hints",
          "ui_descriptor_keys",
          "columns"
        ],
        "title": "PublicationContractModel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Homelab Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/callback": {
      "get": {
        "operationId": "oidc_callback_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oidc Callback"
      }
    },
    "/auth/login": {
      "get": {
        "operationId": "start_oidc_login_auth_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Return To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Oidc Login"
      },
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
...
    "/contracts/publications/{publication_key}": {
      "get": {
        "operationId": "get_publication_contract_contracts_publications__publication_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "publication_key",
            "required": true,
            "schema": {
              "title": "Publication Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationContractModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Publication Contract"
      }
    },
...
  }
}